add concrete agent caller that integrates all parts

+++ src/main/java/it/unibz/butterfly_net/callable_agent/application/CapabilityRegistration.java
@@ -43,7 +43,7 @@ public class CapabilityRegistration {
                             ),
                             "parserInputDescriptors", Set.of(
                                     Map.of(
-                                            "type", "Long",
+                                            "type", "Integer",
                                             "name", "projectId",
                                             "required", true
                                     ),
